app/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
azure-identity==1.14.1
azure-keyvault-secrets==4.7.0
azure-storage-blob==12.19.0
kubernetes==28.1.0
pytest==8.0.0
requests==2.31.0
gitpython==3.1.40